Fix mate-terminal in GTK 3.14
mate-terminal 1.16 uses GTK3.
Its style is defined in gnome-applications.css (TerminalScreen)
and that works with gnome-terminal in LMDE.

It also works fine with mate-terminal 1.16 in Mint 18.x (GTK 3.18).

But it doesn't work at all with mate-terminal 1.16 and GTK 3.14.
Instead, it reads theme_fg_color and theme_base_color.

To work around this issue, we stop using these colors in Mint-X,
and define them only for mate-terminal.